create RetrofitBuilderHandler
