+++ src/main/java/io/github/liuziyuan/retrofit/RetrofitResourceDefinitionRegistry.java
@@ -1,19 +1,25 @@
 package io.github.liuziyuan.retrofit;
 
+import io.github.liuziyuan.retrofit.handler.RetrofitBuilderHandler;
+import io.github.liuziyuan.retrofit.proxy.RetrofitServiceProxyFactory;
+import io.github.liuziyuan.retrofit.resource.RetrofitClientBean;
+import io.github.liuziyuan.retrofit.resource.RetrofitServiceBean;
 import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.beans.factory.support.BeanDefinitionBuilder;
-import org.springframework.beans.factory.support.BeanDefinitionRegistry;
-import org.springframework.beans.factory.support.BeanDefinitionRegistryPostProcessor;
-import org.springframework.beans.factory.support.GenericBeanDefinition;
+import org.springframework.beans.factory.support.*;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
+import retrofit2.Retrofit;
+
+import java.util.List;
+import java.util.Objects;
 
 /**
  * Retrofit Resources Definition and Registry
- * @author liuziyuan
  *
+ * @author liuziyuan
  */
 public class RetrofitResourceDefinitionRegistry implements BeanDefinitionRegistryPostProcessor, ApplicationContextAware {
 
@@ -32,9 +38,14 @@ public class RetrofitResourceDefinitionRegistry implements BeanDefinitionRegistr
             RetrofitResourceContext context = (RetrofitResourceContext) beanFactory.getBean(RetrofitResourceContext.class.getName());
             context.setApplicationContext(applicationContext);
             beanDefinitionRegistry.removeBeanDefinition(RetrofitResourceContext.class.getName());
-            builder = BeanDefinitionBuilder.genericBeanDefinition(RetrofitResourceContext.class, () -> context);
-            GenericBeanDefinition definition = (GenericBeanDefinition) builder.getRawBeanDefinition();
-            beanDefinitionRegistry.registerBeanDefinition(RetrofitResourceContext.class.getName(), definition);
+            List<RetrofitClientBean> retrofitClientBeanList = null;
+            if (!context.getRetrofitClients().isEmpty()) {
+                builder = BeanDefinitionBuilder.genericBeanDefinition(RetrofitResourceContext.class, () -> context);
+                GenericBeanDefinition definition = (GenericBeanDefinition) builder.getRawBeanDefinition();
+                beanDefinitionRegistry.registerBeanDefinition(RetrofitResourceContext.class.getName(), definition);
+                retrofitClientBeanList = context.getRetrofitClients();
+            }
+
         }
     }
