+++ src/main/java/io/github/liuziyuan/retrofit/RetrofitResourceImportDefinitionRegistry.java
@@ -4,10 +4,7 @@ import io.github.liuziyuan.retrofit.annotation.EnableRetrofit;
 import io.github.liuziyuan.retrofit.annotation.RetrofitBuilder;
 import io.github.liuziyuan.retrofit.annotation.RetrofitInterceptor;
 import io.github.liuziyuan.retrofit.extension.UrlOverWriteInterceptor;
-import io.github.liuziyuan.retrofit.handler.CallAdapterFactoryHandler;
-import io.github.liuziyuan.retrofit.handler.ConverterFactoryHandler;
-import io.github.liuziyuan.retrofit.handler.OkHttpClientBuilderHandler;
-import io.github.liuziyuan.retrofit.handler.OkHttpInterceptorHandler;
+import io.github.liuziyuan.retrofit.handler.*;
 import io.github.liuziyuan.retrofit.proxy.RetrofitServiceProxyFactory;
 import io.github.liuziyuan.retrofit.resource.RetrofitClientBean;
 import io.github.liuziyuan.retrofit.resource.RetrofitServiceBean;
@@ -76,7 +73,8 @@ public class RetrofitResourceImportDefinitionRegistry implements ImportBeanDefin
         //registry Retrofit
         for (RetrofitClientBean clientBean : retrofitClientBeanList) {
             builder = BeanDefinitionBuilder.genericBeanDefinition(Retrofit.class, () -> {
-                final Retrofit.Builder retrofitBuilder = getRetrofitBuilder(clientBean);
+                RetrofitBuilderHandler retrofitBuilderHandler = new RetrofitBuilderHandler(clientBean, context);
+                final Retrofit.Builder retrofitBuilder = retrofitBuilderHandler.generate();
                 return retrofitBuilder.build();
             });
             AbstractBeanDefinition definition = builder.getRawBeanDefinition();
@@ -96,85 +94,6 @@ public class RetrofitResourceImportDefinitionRegistry implements ImportBeanDefin
                 registry.registerBeanDefinition(serviceBean.getSelfClazz().getName(), definition);
             }
         }
-    }
-
-    @SneakyThrows
-    private Retrofit.Builder getRetrofitBuilder(RetrofitClientBean clientBean) {
-        Retrofit.Builder builder = new Retrofit.Builder();
-        builder.baseUrl(clientBean.getRealHostUrl());
-        setRetrofitCallAdapterFactory(clientBean, builder);
-        setRetrofitConverterFactory(clientBean, builder);
-        setRetrofitOkHttpClient(clientBean, builder);
-        return builder;
-    }
-
-    private void setRetrofitCallAdapterFactory(RetrofitClientBean clientBean, Retrofit.Builder builder) {
-        final RetrofitBuilder retrofitBuilder = clientBean.getRetrofitBuilder();
-        final List<CallAdapter.Factory> callAdapterFactories = getCallAdapterFactories(retrofitBuilder.addCallAdapterFactory());
-        if (!CollectionUtils.isEmpty(callAdapterFactories)) {
-            callAdapterFactories.forEach(builder::addCallAdapterFactory);
-        }
-    }
-
-    private void setRetrofitConverterFactory(RetrofitClientBean clientBean, Retrofit.Builder builder) {
-        final RetrofitBuilder retrofitBuilder = clientBean.getRetrofitBuilder();
-        final List<Converter.Factory> converterFactories = getConverterFactories(retrofitBuilder.addConverterFactory());
-        if (!CollectionUtils.isEmpty(converterFactories)) {
-            converterFactories.forEach(builder::addConverterFactory);
-        }
-    }
-
-    @SneakyThrows
-    private void setRetrofitOkHttpClient(RetrofitClientBean clientBean, Retrofit.Builder builder) {
-        final RetrofitBuilder retrofitBuilder = clientBean.getRetrofitBuilder();
-        final List<RetrofitInterceptor> interceptors = clientBean.getInterceptors();
-        OkHttpClient.Builder okHttpClientBuilder;
-        OkHttpClientBuilderHandler okHttpClientBuilderHandler;
-        if (retrofitBuilder.client() != null) {
-            okHttpClientBuilderHandler = new OkHttpClientBuilderHandler(retrofitBuilder.client());
-            okHttpClientBuilder = okHttpClientBuilderHandler.generate();
-        } else {
-            okHttpClientBuilder = new OkHttpClient.Builder();
-        }
-        okHttpClientBuilder.addInterceptor(new UrlOverWriteInterceptor(context));
-        final List<Interceptor> okHttpInterceptors = getOkHttpInterceptor(interceptors);
-
-        okHttpInterceptors.forEach(okHttpClientBuilder::addInterceptor);
-        builder.client(okHttpClientBuilder.build());
-    }
-
-    @SneakyThrows
-    private List<Interceptor> getOkHttpInterceptor(List<RetrofitInterceptor> interceptors) {
-        List<Interceptor> interceptorList = new ArrayList<>();
-        OkHttpInterceptorHandler okHttpInterceptorHandler;
-        for (RetrofitInterceptor interceptor : interceptors) {
-            okHttpInterceptorHandler = new OkHttpInterceptorHandler(interceptor.handler(), context);
-            final Interceptor generateInterceptor = okHttpInterceptorHandler.generate();
-            interceptorList.add(generateInterceptor);
-        }
-        return interceptorList;
-    }
-
-    @SneakyThrows
-    private List<CallAdapter.Factory> getCallAdapterFactories(Class<? extends CallAdapter.Factory>[] callAdapterFactoryClasses) {
-        List<CallAdapter.Factory> callAdapterFactories = new ArrayList<>();
-        CallAdapterFactoryHandler adapterFactoryHandler;
-        for (Class<? extends CallAdapter.Factory> callAdapterFactoryClass : callAdapterFactoryClasses) {
-            adapterFactoryHandler = new CallAdapterFactoryHandler(callAdapterFactoryClass);
-            callAdapterFactories.add(adapterFactoryHandler.generate());
-        }
-        return callAdapterFactories;
-    }
-
-    @SneakyThrows
-    private List<Converter.Factory> getConverterFactories(Class<? extends Converter.Factory>[] converterFactoryClasses) {
-        List<Converter.Factory> converterFactories = new ArrayList<>();
-        ConverterFactoryHandler converterFactoryHandler;
-        for (Class<? extends Converter.Factory> converterFactoryClass : converterFactoryClasses) {
-            converterFactoryHandler = new ConverterFactoryHandler(converterFactoryClass);
-            converterFactories.add(converterFactoryHandler.generate());
-        }
-        return converterFactories;
     }
 
     @Override
